Refactor audio (#9)
This refactoring dramatically improves the sound generator. It supports all four
channels, have some debugging UI and some basic support for recording audio
to disk. There's some basic support for dynamic playback rate, but more remains
to be done. Right now it assumes the frame rate is fixed at 60 fps, and even then
it's not perfect. But I think this is a good point to submit the work and start running
test roms to nail the last pieces of the sound generation.