Fix obscure behavior when reading wave memory while playing (#19)
This commit fixes some obscure behavior that should likely have very little impact on
how well the emulator can run games and other software. But I want to pass all Blargg
tests, so this had to be done.

For CGB, the last played sample is always returned when wave memory is read if the
wave channel is playing.

For DMG it's the same, except that it only does so for 2 cycles (?). After that 0xFF
is returned.