Reset frame sequencer step when APU is powered on (#20)
This commit adds a counter to the APU instead of deriving
the current frame sequencer step from the DIV counter. This
is done so that the step can be reset when the APU is powered
on.

When the APU is powered on the frame sequencer should be
reset so that its *next* step is step 0.

This makes Blargg audio test `07-len sweep period sync` pass.
It is no longer skipped in the CI test suite. For some reason
`11-regs after power` was also skipped before, but it runs
and pass the test, so I've enabled that one as well.