PPU refactor (#29)
Initial intent was to implement the full FIFO. But after finding that it's still
pretty undocumented and only enables one single released game and a
few demos to work flawlessly, while giving a substantial performance
drop in the general case, I decided that I better improve the existing
PPU implementation. In the end, I decided to rewrite the PPU code
from scratch.

Neither the old nor the new implementation is anywhere close to pass
the ACID2 test, and the FIFO implementation just barely manage to
show the Nintendo logo at boot. But I recon this could be a good time
to merge the work anyway. The previous PPU implementation lcd.rs,
and the fifo implementation (ppu_fifo) will be dropped soon.