Pass dmg-acid2 test (lots of fixes in PPU and UI) (#30)
With this commit, there's a new implementation of the PPU which successfully pass
the `dmg-acid2` test. But while the new PPU code is an improved implementation,
there are many regressions, for example in sprites/OAM objects. They don't show
up where they should, not even in a simple game as Tetris. But that should be
quite easy to fix.

- Revised UI with a list of windows on the left and debugger buttons at the top
- New Pixbuf type that handles image buffers and textures
- Window that shows the tile map and tile data
- New PPU implementation from scratch!
- Add dmg-acid2 test to CI (not perfect, does not show image, just pass/fail status)
- Upgrade many dependencies